+++ Cargo.toml
@@ -49,6 +49,7 @@ parquet = { version = "51.0.0", features = ["async"] }
 arrow = { version = "51.0.0", features = ["prettyprint"] }
 async-stream = "0.3.5"
 thiserror = "1.0.58"
+strum = { version = "0.26.2", features = ["derive"] }
 
 [dev-dependencies]
 float_eq = "1.0.1"
